ref:wip: refactoring turtle definition for 2D L-Systems
A turtle was a {position, orientation} pair, which was incorrect in 3D as
rotation order matter. (`orientation` was an angle.)

Now a turtle is a {position, direction} pair. On each `turn A`, the new
`direction` is calculated using a rotation matrix R(A). Afterward, moving
`forward N` is just `position` = `position` + N * `direction`.